refactor(visualise): move utility functions
- Moved utils/display.py to commands/visualise/utils/display.py.
- Modified import references to match new path.
